bsale/bSaleServer/requirements.txt
asgiref==3.5.2
astroid==2.4.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2022.4
sqlparse==0.4.3
toml==0.10.2
wrapt==1.12.1
